requirements.txt
diffusers==0.19.1
einops==0.6.1
gradio==3.32.0
kornia==0.6.8
numpy==1.23.5
omegaconf==2.2.2
opencv_python_headless==4.6.0.66
Pillow==9.4.0
pytorch_lightning==1.9.3
torch==2.0.0
torchvision==0.15.1
tqdm==4.65.0
transformers==4.29.2
accelerate==0.19.0
pyequilib @ https://github.com/zhuqiangLu/equilib.git
